Separated db into separate module because I get the feeling that will be helpful in the future
