+++ db/go.mod
@@ -1,0 +1,5 @@
+module db
+
+go 1.19
+
+require github.com/andersfylling/snowflake/v5 v5.0.1
