+++ go.mod
@@ -2,12 +2,10 @@ module github.com/10Nates/tras
 
 go 1.19
 
-require (
-	github.com/andersfylling/disgord v0.36.0
-	github.com/andersfylling/snowflake/v5 v5.0.1
-)
+require github.com/andersfylling/disgord v0.36.0
 
 require (
+	github.com/andersfylling/snowflake/v5 v5.0.1 // indirect
 	github.com/klauspost/compress v1.15.11 // indirect
 	go.uber.org/atomic v1.10.0 // indirect
 	golang.org/x/crypto v0.1.0 // indirect
