+++ go.work
@@ -1,0 +1,4 @@
+go 1.19
+
+use ./
+use ./db
